requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.3.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.13.1
knockknock==0.1.8.1
lxml==4.9.2
matrix-client==0.4.0
more-itertools==9.1.0
multidict==6.0.4
outcome==1.2.0
packaging==23.0
premailer==3.10.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
SecretStorage==3.3.3
selenium==4.8.3
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
twilio==8.0.0
urllib3==1.26.15
webdriver-manager==3.8.5
wsproto==1.2.0
yagmail==0.15.293
yarl==1.8.2
zipp==3.15.0
